Update MatchTimer.java
I'll post this on the forum aswell, not sure if this is going into a black hole or someone will actually notice it.

This is a working timer, with glitches. Even scheduledatfixed rate doesen't seem as accurate as I thought it would be, but I have yet to do some more testing. When screen dims the timer stops, but I think this is a simple wakelock permission thing.   I created 2 layouts for testing, and this activity was called from my main activity. Would be pointless putting them up, this may have to be changed a little for Colm or Carl. The code is remarkably simple for the time I spent on it. Am trying to get Pause to work but there are a few complications using it with timertask, think  the timer has to be cancelled and restarted using but havn't looked into it fully yet.