Eliminated 6 characters
The other one is with scheduledatfixedrate. Keep both somehow if possible, and test the 2. schedulefixedrate does it by ensuring the integrity longterm, so if a beep was 1.3 seconds late for whatever reason the other beep will occur .2 seconds later. With schedule it would occur 1.5 seconds later regardless of how late the initial beep was. It is less likely to be noticed but actually less accurate.

See here that there is another class within the oncreate. If we go with this timer it will create problems for you, might have to be put into its own class file and result in a million other complications. Finally, you have to create a raw folder with a music file or else your own folder and file for r.raw.beep.