Same thing except with pause function to cancel the timer.
The Resume function has to be kept in its own class file. Its the same code repeated, and if kept in this oncreate, increases the number of 'beeps', as it seems to run through the code twice. Might be a better way around it but for now should do the trick.